requirements.txt
alembic==1.13.3
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-telegram-bot==21.6
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
